refactor sse broadcast to shared module
