+++ web/src/routes/api/experiments/[slug]/metrics/batch/+server.ts
@@ -1,6 +1,6 @@
 import { json, type RequestEvent } from "@sveltejs/kit";
 import type { RequestHandler } from "./$types";
-import { _broadcastMetric } from "../../../../metrics/[experimentId]/stream/+server";
+import { broadcastMetric } from "$lib/server/broadcast";
 import type { Json } from "$lib/server/database.types";
 import type { Metric } from "$lib/types";
 
@@ -61,7 +61,7 @@ export const POST: RequestHandler = async ({ request, params, locals }) => {
     await locals.dbClient.batchCreateMetric(finalMetrics);
 
     for (const m of metrics) {
-      _broadcastMetric(experimentId, JSON.stringify(m));
+      broadcastMetric(experimentId, JSON.stringify(m));
     }
 
     return json({
